Organized and Renamed Files
Cleaned up the code and split the functions into appropriate files and cleared any non-useful files
