+++ output.mc
@@ -1,12 +1,19 @@
-0x00000000 0x00200093 , addi 1 0 2 # 00000000001000000000000010010011
-0x00000004 0x00200113 , addi 2 0 2 # 00000000001000000000000100010011
-0x00000008 0x00208863 , beq 1, 2, label1 # 00000000001000001000100001100011
-0x0000000c 0x00419863 , bne 3, 4, label2 # 00000000010000011001100001100011
-0x00000010 0x0062C863 , blt 5, 6, label3 # 00000000011000101100100001100011
-0x00000014 0x0083D863 , bge 7, 8, label4 # 00000000100000111101100001100011
-0x00000018 0x00B504B3 , add 9, 10, 11 # 00000000101101010000010010110011
-0x0000001c 0x40E68633 , sub 12, 13, 14 # 01000000111001101000011000110011
-0x00000020 0x011877B3 , and 15, 16, 17 # 00000001000110000111011110110011
-0x00000024 0x0149E933 , or 18, 19, 20 # 00000001010010011110100100110011
-0x00000028 0x017B4AB3 , xor 21, 22, 23 # 00000001011110110100101010110011
-0x0000002c END_OF_TEXT
+.data
+0x10000000 0x03
+0x10000001 0x14
+0x10000002 0x000000ab
+0x10000006 0x03
+0x10000007 0x61
+0x10000008 0x00
+.text
+0x00000000 0x00A10083  lb x1, x2, 10  # 0000011 - 000 - NULL - 00001 - 00010 - NULL - 000000001010
+0x00000004 0x00020113  addi x2, x0, 2  # 0010011 - 000 - NULL - 00010 - 0 - NULL - 000000000010
+0x00000008 0x02208063  beq x1, x2, 16  # 1100011 - 000 - NULL - NULL - 00001 - 00010 - 000000010000
+0x0000000c 0x02419063  bne x3, x4, 16  # 1100011 - 001 - NULL - NULL - 00011 - 00100 - 000000010000
+0x00000010 0x0262C063  blt x5, x6, 16  # 1100011 - 100 - NULL - NULL - 00101 - 00110 - 000000010000
+0x00000014 0x018000EF  jal x1, 12  # 1101111 - NULL - NULL - 00001 - NULL - NULL - 00000000000000001100
+0x00000018 0x00B504B3  add x9, x10, x11  # 0110011 - 000 - 0000000 - 01001 - 01010 - 01011 - NULL
+0x0000001c 0x40E68633  sub x12, x13, x14  # 0110011 - 000 - 0100000 - 01100 - 01101 - 01110 - NULL
+0x00000020 0x011877B3  and x15, x16, x17  # 0110011 - 111 - 0000000 - 01111 - 10000 - 10001 - NULL
+0x00000024 0x0149E933  or x18, x19, x20  # 0110011 - 110 - 0000000 - 10010 - 10011 - 10100 - NULL
+0x00000028 0x017B4AB3  xor x21, x22, x23  # 0110011 - 100 - 0000000 - 10101 - 10110 - 10111 - NULL
